+++ tests/integ/scala/test_dataframe_aggregate_suite.py
@@ -47,6 +47,10 @@ from snowflake.snowpark.functions import (
 from tests.utils import IS_IN_STORED_PROC, TestData, Utils
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: pivot not supported",
+)
 def test_pivot(session):
     Utils.check_answer(
         TestData.monthly_sales(session)
@@ -68,6 +72,10 @@ def test_pivot(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: pivot not supported",
+)
 def test_group_by_pivot(session):
     Utils.check_answer(
         TestData.monthly_sales_with_team(session)
@@ -102,6 +110,10 @@ def test_group_by_pivot(session):
         ).agg([sum(col("amount")), avg(col("amount"))])
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: SNOW-1370114 pivot should raise not implemented error but get AttributeError: DataFrame object has no attribute queries",
+)
 def test_group_by_pivot_dynamic_any(session, caplog):
     Utils.check_answer(
         TestData.monthly_sales_with_team(session)
@@ -137,6 +149,10 @@ def test_group_by_pivot_dynamic_any(session, caplog):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: SNOW-1370114 pivot should raise not implemented error but get AttributeError: DataFrame object has no attribute queries",
+)
 def test_group_by_pivot_dynamic_subquery(session):
     src = TestData.monthly_sales(session)
     subquery_df = src.select(col("month")).filter(col("month") == "JAN")
@@ -162,6 +178,10 @@ def test_group_by_pivot_dynamic_subquery(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: pivot not supported",
+)
 def test_join_on_pivot(session):
     df1 = (
         TestData.monthly_sales(session)
@@ -182,6 +202,10 @@ def test_join_on_pivot(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: pivot not supported",
+)
 def test_pivot_on_join(session):
     df = session.create_dataframe([[1, "One"], [2, "Two"]]).to_df("empid", "name")
 
@@ -202,6 +226,10 @@ def test_pivot_on_join(session):
 # TODO (SNOW-916206)  If the source is a temp table with inlined data, then we need to validate that
 # pivot will materialize the data before executing pivot, otherwise would fail with not finding the
 # data when doing a later schema call.
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: SNOW-1370114 pivot should raise not implemented error but get AttributeError: DataFrame object has no attribute queries",
+)
 @pytest.mark.skipif(IS_IN_STORED_PROC, reason="pivot does not work in stored proc")
 def test_pivot_dynamic_any_with_temp_table_inlined_data(session):
     original_df = session.create_dataframe(
@@ -222,6 +250,10 @@ def test_pivot_dynamic_any_with_temp_table_inlined_data(session):
     assert pivot_op_df.count() == 1
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: SNOW-1370114 pivot should raise not implemented error but get AttributeError: DataFrame object has no attribute queries",
+)
 def test_pivot_dynamic_any(session):
     Utils.check_answer(
         TestData.monthly_sales(session)
@@ -236,6 +268,10 @@ def test_pivot_dynamic_any(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: SNOW-1370114 pivot should raise not implemented error but get AttributeError: DataFrame object has no attribute queries",
+)
 def test_pivot_dynamic_subquery(session):
     src = TestData.monthly_sales(session)
     subquery_df = src.select(col("month")).filter(col("month") == "JAN")
@@ -302,6 +338,10 @@ def test_pivot_dynamic_subquery_with_bad_subquery(session):
     assert "Pivot subquery must select single column" in str(ex_info.value)
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: pivot not supported",
+)
 def test_pivot_default_on_none(session, caplog):
     class MonthlySales(NamedTuple):
         empid: int
@@ -399,6 +439,10 @@ def test_group_by(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: SNOW-977749 grouping by grouping sets not supported",
+)
 def test_group_by_grouping_sets(session):
     result = (
         TestData.nurse(session)
@@ -809,6 +853,10 @@ def test_count(session):
     ).collect() == [Row(6, 6.0)]
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: stddev not supported",
+)
 def test_stddev(session):
     test_data_dev = sqrt(4 / 5)
 
@@ -828,6 +876,10 @@ def test_stddev(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: variance not supported",
+)
 def test_sn_moments(session):
     test_data2 = TestData.test_data2(session)
     var = test_data2.agg(variance(col("a")))
@@ -860,6 +912,10 @@ def test_sn_moments(session):
     Utils.check_answer(kurtosis_, agg_kurtosis_result[0])
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: stddev not supportedg",
+)
 def test_sn_zero_moments(session):
     input = session.create_dataframe([[1, 2]]).to_df("a", "b")
     Utils.check_answer(
@@ -895,6 +951,10 @@ def test_sn_zero_moments(session):
     )
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: variance not supported",
+)
 def test_sn_null_moments(session):
     empty_table_data = session.create_dataframe([[]]).to_df("a")
 
@@ -926,12 +986,9 @@ def test_sn_null_moments(session):
 
 
 def test_decimal_sum_over_window_should_work(session):
-    assert session.sql(
-        "select sum(a) over () from values (1.0), (2.0), (3.0) T(a)"
-    ).collect() == [Row(6.0), Row(6.0), Row(6.0)]
-    assert session.sql(
-        "select avg(a) over () from values (1.0), (2.0), (3.0) T(a)"
-    ).collect() == [Row(2.0), Row(2.0), Row(2.0)]
+    df = session.create_dataframe([1.0, 2.0, 3.0], schema=["a"])
+    assert df.select(sum("a").over()).collect() == [Row(6.0), Row(6.0), Row(6.0)]
+    assert df.select(avg("a").over()).collect() == [Row(2.0), Row(2.0), Row(2.0)]
 
 
 @pytest.mark.localtest
@@ -952,6 +1009,11 @@ def test_ints_in_agg_exprs_are_taken_as_groupby_ordinal(session):
     ).collect() == [Row(3, 4, 6, 1, 3), Row(3, 4, 6, 2, 6)]
 
 
+@pytest.mark.xfail(
+    "config.getvalue('local_testing_mode')",
+    reason="SQL query not supported",
+    run=False,
+)
 def test_ints_in_agg_exprs_are_taken_as_groupby_ordinal_sql(session):
 
     testdata2str = "(SELECT * FROM VALUES (1,1),(1,2),(2,1),(2,2),(3,1),(3,2) T(a, b) )"
@@ -1020,6 +1082,10 @@ def test_distinct_and_unionall(session):
     assert res == [Row("one"), Row("one")]
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: count_if not suppored in snowpark python",
+)
 def test_count_if(session):
     temp_view_name = Utils.random_name_for_temp_object(TempObjectType.VIEW)
     session.create_dataframe(
@@ -1171,6 +1237,10 @@ def test_zero_count(session):
     ]
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: stddev not supported",
+)
 def test_zero_stddev(session):
     df = session.create_dataframe([[]]).to_df(["a"])
     assert df.agg(
@@ -1221,6 +1291,10 @@ def test_listagg(session):
     assert len(result) == 4
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: aggregate expression within group not supported",
+)
 def test_listagg_within_group(session):
     df = session.create_dataframe(
         [