+++ tests/integ/test_df_to_pandas.py
@@ -8,6 +8,7 @@ from typing import Iterator
 import pytest
 
 try:
+    import numpy as np
     import pandas as pd
     from pandas import DataFrame as PandasDF, Series as PandasSeries
     from pandas.testing import assert_frame_equal
@@ -20,10 +21,35 @@ except ImportError:
     pytest.skip("pyarrow is not available", allow_module_level=True)
 
 
+import datetime
+import decimal
+
+import pytest
+
 from snowflake.snowpark._internal.utils import TempObjectType
 from snowflake.snowpark.exceptions import SnowparkFetchDataException
-from snowflake.snowpark.functions import col, to_timestamp
-from snowflake.snowpark.types import DecimalType, IntegerType
+from snowflake.snowpark.functions import col, div0, round, to_timestamp
+from snowflake.snowpark.types import (
+    ArrayType,
+    BinaryType,
+    BooleanType,
+    ByteType,
+    DateType,
+    DecimalType,
+    DoubleType,
+    FloatType,
+    IntegerType,
+    LongType,
+    MapType,
+    NullType,
+    ShortType,
+    StringType,
+    StructField,
+    StructType,
+    TimestampType,
+    TimeType,
+    VariantType,
+)
 from tests.utils import IS_IN_STORED_PROC, Utils
 
 
@@ -112,24 +138,29 @@ def test_to_pandas_cast_integer(session, to_pandas_api, local_testing_mode):
         assert str(timestamp_pandas_df.dtypes[0]) == "datetime64[ns]"
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="BUG: dtype and value mismatch",
+)
 def test_to_pandas_precision_for_number_38_0(session):
     # Assert that we try to fit into int64 when possible and keep precision
-    df = session.sql(
-        """
-    SELECT
-        CAST(COLUMN1 as NUMBER(38,0)) AS A,
-        CAST(COLUMN2 as NUMBER(18,0)) AS B
-    FROM VALUES
-        (1111111111111111111, 222222222222222222),
-        (3333333333333333333, 444444444444444444),
-        (5555555555555555555, 666666666666666666),
-        (7777777777777777777, 888888888888888888),
-        (9223372036854775807, 111111111111111111),
-        (2222222222222222222, 333333333333333333),
-        (4444444444444444444, 555555555555555555),
-        (6666666666666666666, 777777777777777777),
-        (-9223372036854775808, 999999999999999999)
-        """
+
+    df = session.create_dataframe(
+        [
+            [1111111111111111111, 222222222222222222],
+            [3333333333333333333, 444444444444444444],
+            [5555555555555555555, 666666666666666666],
+            [7777777777777777777, 888888888888888888],
+            [9223372036854775807, 111111111111111111],
+            [2222222222222222222, 333333333333333333],
+            [4444444444444444444, 555555555555555555],
+            [6666666666666666666, 777777777777777777],
+            [-9223372036854775808, 999999999999999999],
+        ],
+        schema=["A", "B"],
+    ).select(
+        col("A").cast(DecimalType(38, 0)).alias("A"),
+        col("B").cast(DecimalType(18, 0)).alias("B"),
     )
 
     pdf = df.to_pandas()
@@ -141,19 +172,20 @@ def test_to_pandas_precision_for_number_38_0(session):
     assert pdf["A"].min() == -9223372036854775808
 
 
+@pytest.mark.skipif(
+    "config.getvalue('local_testing_mode')",
+    reason="FEAT: div0 and round functions not supported",
+)
 def test_to_pandas_precision_for_non_zero_scale(session):
-    df = session.sql(
-        """
-        SELECT
-            num1,
-            num2,
-            DIV0(num1, num2) AS A,
-            DIV0(CAST(num1 AS INTEGER), CAST(num2 AS INTEGER)) AS B,
-            ROUND(B, 2) as C
-        FROM (VALUES
-            (1, 11)
-        ) X(num1, num2);
-        """
+
+    df = session.create_dataframe([[1, 11]], schema=["num1", "num2"]).select(
+        col("num1"),
+        col("num2"),
+        div0(col("num1"), col("num2")).alias("A"),
+        div0(col("num1").cast(IntegerType()), col("num2").cast(IntegerType())).alias(
+            "B"
+        ),
+        round(col("B"), 2).alias("C"),
     )
 
     pdf = df.to_pandas()
@@ -163,6 +195,11 @@ def test_to_pandas_precision_for_non_zero_scale(session):
     assert pdf["C"].dtype == "float64"
 
 
+@pytest.mark.xfail(
+    "config.getvalue('local_testing_mode')",
+    reason="SQL query not supported",
+    run=False,
+)
 def test_to_pandas_non_select(session):
     # `with ... select ...` is also a SELECT statement
     isinstance(session.sql("select 1").to_pandas(), PandasDF)
@@ -226,3 +263,124 @@ def test_to_pandas_batches(session, local_testing_mode):
     for df_batch in df.to_pandas_batches():
         assert_frame_equal(df_batch, entire_pandas_df.iloc[: len(df_batch)])
         break
+
+
+@pytest.mark.localtest
+def test_df_to_pandas_df(session):
+    df = session.create_dataframe(
+        [
+            [
+                1,
+                1234567890,
+                True,
+                1.23,
+                "abc",
+                b"abc",
+                datetime.datetime(
+                    year=2023, month=10, day=30, hour=12, minute=12, second=12
+                ),
+            ]
+        ],
+        schema=[
+            "aaa",
+            "BBB",
+            "cCc",
+            "DdD",
+            "e e",
+            "ff ",
+            " gg",
+        ],
+    )
+
+    to_compare_df = pd.DataFrame(
+        {
+            "AAA": pd.Series([1], dtype=np.int64),
+            "BBB": pd.Series([1234567890], dtype=np.int64),
+            "CCC": pd.Series([True]),
+            "DDD": pd.Series([1.23]),
+            "e e": pd.Series(["abc"]),
+            "ff ": pd.Series([b"abc"]),
+            " gg": pd.Series(
+                [
+                    datetime.datetime(
+                        year=2023, month=10, day=30, hour=12, minute=12, second=12
+                    )
+                ]
+            ),
+        }
+    )
+
+    # assert_frame_equal also checks dtype
+    assert_frame_equal(df.to_pandas(), to_compare_df)
+    assert_frame_equal(list(df.to_pandas_batches())[0], to_compare_df)
+
+    # check snowflake types explicitly
+    df = session.create_dataframe(
+        data=[
+            [
+                [1, 2, 3, 4],
+                b"123",
+                True,
+                1,
+                datetime.date(year=2023, month=10, day=30),
+                decimal.Decimal(1),
+                1.23,
+                1.23,
+                100,
+                100,
+                None,
+                100,
+                "abc",
+                datetime.datetime(2023, 10, 30, 12, 12, 12),
+                datetime.time(12, 12, 12),
+                {"a": "b"},
+                {"a": "b"},
+            ],
+        ],
+        schema=StructType(
+            [
+                StructField("a", ArrayType()),
+                StructField("b", BinaryType()),
+                StructField("c", BooleanType()),
+                StructField("d", ByteType()),
+                StructField("e", DateType()),
+                StructField("f", DecimalType()),
+                StructField("g", DoubleType()),
+                StructField("h", FloatType()),
+                StructField("i", IntegerType()),
+                StructField("j", LongType()),
+                StructField("k", NullType()),
+                StructField("l", ShortType()),
+                StructField("m", StringType()),
+                StructField("n", TimestampType()),
+                StructField("o", TimeType()),
+                StructField("p", VariantType()),
+                StructField("q", MapType(StringType(), StringType())),
+            ]
+        ),
+    )
+
+    pandas_df = pd.DataFrame(
+        {
+            "A": pd.Series(["[\n  1,\n  2,\n  3,\n  4\n]"], dtype=object),
+            "B": pd.Series([b"123"], dtype=object),
+            "C": pd.Series([True], dtype=bool),
+            "D": pd.Series([1], dtype=np.int64),
+            "E": pd.Series([datetime.date(year=2023, month=10, day=30)], dtype=object),
+            "F": pd.Series([decimal.Decimal(1)], dtype=np.int64),
+            "G": pd.Series([1.23], dtype=np.float64),
+            "H": pd.Series([1.23], dtype=np.float64),
+            "I": pd.Series([100], dtype=np.int64),
+            "J": pd.Series([100], dtype=np.int64),
+            "K": pd.Series([None], dtype=object),
+            "L": pd.Series([100], dtype=np.int64),
+            "M": pd.Series(["abc"], dtype=object),
+            "N": pd.Series(
+                [datetime.datetime(2023, 10, 30, 12, 12, 12)], dtype="datetime64[ns]"
+            ),
+            "O": pd.Series([datetime.time(12, 12, 12)], dtype=object),
+            "P": pd.Series(['{\n  "a": "b"\n}'], dtype=object),
+            "Q": pd.Series(['{\n  "a": "b"\n}'], dtype=object),
+        }
+    )
+    assert_frame_equal(df.to_pandas(), pandas_df)